refactor(controller): move bulk of logic into the resource service
